mcp-server/requirements.txt
mcp
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
python-dotenv==1.0.1
pydantic==2.7.4
